LabExamples/Lab06_Carpisma_Algilama/pyproject.toml
[project]
name = "lab06-carpisma-algilama"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["pygame-ce>=2.5.7"]
